Fix: Item hover events no longer support the "value" field
On MC 1.21.5, we observe an error when we try to send item data via the
"value" field of hover events.

BungeeCord chat does not yet support component data in item hover events.
So on MC 1.21.5+, we only show the item type in hover events for now.
For previous server versions we still use the "value" field with the full
item data as before.

Internal changes to HoverEventText representation.