Fix: Mark shopkeepers as dirty when a property value is found missing
As before, we log a warning and then fall back to the default value for
missing property values. This is for example often the case when new
properties are added.

However, previously, we would not mark the shopkeeper as dirty in this case
since the default value also equals the initial property value. One
consequence of this was that the warning about missing property values
would be repeated over and over during subsequent plugin reloads.