Add support for Paper 1.21.8
Internally, this reuses the same compatibility implementation as for
1.21.7.

A single CompatVersion can now support multiple server versions: This is
not strictly required for Spigot as long as the mappings versions does not
change. But for Paper, we have no mappings version, so we must manually add
a mapping for each new Minecraft version.